:sparkles: move Advertise dir enums inside module in Modules\Advertise\Enums dir
