--- app/Enums/Advertisement/Sort.php
@@ -1,15 +1,0 @@
-<?php
-
-namespace App\Enums\Advertisement;
-
-use App\Enums\Concerns\EnumDataListTrait;
-
-enum Sort: string
-{
-    use EnumDataListTrait;
-
-    case PRICE_ASC = 'price_asc';
-    case PRICE_DESC = 'price_desc';
-    case NEWEST = 'newest';
-    case OLDEST = 'oldest';
-}
