+++ database/seeders/UserBatchSeeder.php
@@ -4,13 +4,13 @@ declare(strict_types=1);
 
 namespace Database\Seeders;
 
-use App\Enums\Advertisement\AdvertisementStatus;
-use App\Enums\Advertisement\AdvertisementType;
 use App\Models\Geo\City;
 use App\Models\User;
 use DB;
 use Hash;
 use Illuminate\Database\Seeder;
+use Modules\Advertise\Enums\AdvertisementStatus;
+use Modules\Advertise\Enums\AdvertisementType;
 use Modules\Advertise\Models\Advertisement;
 use Modules\Advertise\Models\Category;
 use Schema;
